directory and module restructure
