+++ src/dapper/config/met_source_params.py
@@ -1,0 +1,6 @@
+# Declarative knobs per source. Start with ERA5; others can copy the shape.
+
+ERA5 = dict(
+    id_fallbacks=["gid","pids","pid","id"],
+    note="ERA5-Land hourly via GEE",
+)
